+++ vktrace/vktrace_layer/vktrace_lib_pageguardmappedmemory.h
@@ -125,8 +125,8 @@ typedef class PageGuardMappedMemory {
     /// VkDeviceSize RangeOffset, RangeSize, only consider the block which is in the range which start from RangeOffset and size is
     /// RangeSize, if RangeOffset<0, consider whole mapped memory
     /// return the amount of changed blocks.
-    size_t getChangedBlockInfo(VkDeviceSize RangeOffset, VkDeviceSize RangeSize, size_t *pdwSaveSize, size_t *pInfoSize, PBYTE pData,
-                              size_t DataOffset, int useWhich = BLOCK_FLAG_ARRAY_CHANGED);
+    uint64_t getChangedBlockInfo(VkDeviceSize RangeOffset, VkDeviceSize RangeSize, uint64_t *pdwSaveSize, uint64_t *pInfoSize,
+                                 PBYTE pData, uint64_t DataOffset, int useWhich = BLOCK_FLAG_ARRAY_CHANGED);
 
     /// return: if memory already changed;
     ///        evenif no change to mmeory, it will still allocate memory for info array which only include one
