+++ vktrace/vktrace_layer/vktrace_lib_pagestatusarray.cpp
@@ -31,16 +31,16 @@ PageStatusArray::PageStatusArray(uint64_t pageCount) {
     ByteCount = (pageCount % PAGE_FLAG_AMOUNT_PER_BYTE) ? (pageCount / PAGE_FLAG_AMOUNT_PER_BYTE) + 1
                                                         : pageCount / PAGE_FLAG_AMOUNT_PER_BYTE;
 
-    pChangedArray[0] = new uint8_t[ByteCount];
+    pChangedArray[0] = new uint8_t[(size_t)ByteCount];
     assert(pChangedArray[0]);
 
-    pChangedArray[1] = new uint8_t[ByteCount];
+    pChangedArray[1] = new uint8_t[(size_t)ByteCount];
     assert(pChangedArray[1]);
 
-    pReadArray[0] = new uint8_t[ByteCount];
+    pReadArray[0] = new uint8_t[(size_t)ByteCount];
     assert(pReadArray[0]);
 
-    pReadArray[1] = new uint8_t[ByteCount];
+    pReadArray[1] = new uint8_t[(size_t)ByteCount];
     assert(pReadArray[1]);
 
     activeChangesArray = pChangedArray[0];
@@ -48,7 +48,7 @@ PageStatusArray::PageStatusArray(uint64_t pageCount) {
     activeReadArray = pReadArray[0];
     capturedReadArray = pReadArray[1];
 
-    firstTimeLoadArray = new uint8_t[ByteCount];
+    firstTimeLoadArray = new uint8_t[(size_t)ByteCount];
     assert(firstTimeLoadArray);
 
     clearAll();
@@ -141,9 +141,9 @@ void PageStatusArray::backupChangedArray() { toggleChangedArray(); }
 void PageStatusArray::backupReadArray() { toggleReadArray(); }
 
 void PageStatusArray::clearAll() {
-    memset(activeChangesArray, 0, ByteCount);
-    memset(capturedChangesArray, 0, ByteCount);
-    memset(activeReadArray, 0, ByteCount);
-    memset(capturedReadArray, 0, ByteCount);
-    memset(firstTimeLoadArray, 0, ByteCount);
+    memset(activeChangesArray, 0, (size_t)ByteCount);
+    memset(capturedChangesArray, 0, (size_t)ByteCount);
+    memset(activeReadArray, 0, (size_t)ByteCount);
+    memset(capturedReadArray, 0, (size_t)ByteCount);
+    memset(firstTimeLoadArray, 0, (size_t)ByteCount);
 }
